Policy

+++ app/Http/Controllers/FakultasController.php
@@ -31,6 +31,12 @@ class FakultasController extends Controller
      */
     public function store(Request $request)
     {
+        // cek apakah user memiliki hak akses untuk membuat fakultas
+        if ($request->user()->cannot('create', Fakultas::class)) {
+            abort(403);
+        }
+
+        
         // validasi input
         $input = $request->validate([
             'nama' => 'required|unique:fakultas',
@@ -70,6 +76,13 @@ class FakultasController extends Controller
     public function update(Request $request, $fakultas)
     {
         $fakultas = Fakultas::findOrFail($fakultas);
+
+        // cek apakah user memiliki hak akses untuk mengupdate fakultas
+        if ($request->user()->cannot('update', $fakultas)) {
+            abort(403);
+        }
+        
+        
         // validasi input
         $input = $request->validate([
             'nama' => 'required',
@@ -86,10 +99,15 @@ class FakultasController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($fakultas)
+    public function destroy(Request $request, $fakultas)
     {
         $fakultas = Fakultas::findOrFail($fakultas);
         // dd($fakultas);
+
+        if ($request->user()->cannot('delete', $fakultas)) {
+            abort(403);
+        }
+
 
         //hapus data fakultas
         $fakultas->delete();
